[media] build / test : file chunk setup for integration test in docker compose
