+++ services/media/infra/run_integration_test
@@ -1,5 +1,3 @@
-python3 -m media.task_loader  FilechunkSetup  ${ITEST_APPCFG_PATH};
-
 
 make itest_app_server -C ./build
